Expect the kernel of the node image to be at /kernel
Before, the kernel path on the nfs share was specified as a custom
dhcp option, together with the image filesystem path.
It is handy to consider that node images should provide all what is
needed, i.e. both the filesystem and the kernel. Thus, we now
consider that the kernel should be available at path /kernel of the
filesystem image.