Let a new node register itself to the walt server.
DHCPd will not send a default nfs path when a node
boots up for the 1st time.
This way, the new node can detect that it is unknown
from the walt server and that it should register first.
Registration is done by connecting to the walt TCP server
and sending a request with appropriate information about
the node (type, ip, mac).