nfs mount watchdog: added timeout on 'ls' command
The NFS mount watchdog is responsible for detecting that the NFS
root was lost. In this case the node should reboot.
This was working when retarting the service walt-server on the server.
But when rebooting the server (i.e. 'init 6'), for some reason it was
not working: the nodes would hang instead of rebooting. I suspected
the 'ls' command of the watchdog was hanging instead of failing in this
case, so I added a timeout on it. Now rebooting the server works as
expected too.